Refactored task naming to make my acr resolution task be the acr task and the previous idea the hazen custom task. The concept here is to differentiate from the tasks that try to meet ACR specifications. Added labels to the ACR Spatial Resolution Task report.
